# Conflicts:
#	Kooboo.App/App.config
#	Kooboo.Data/AppSettings.cs
#	Kooboo.Lib/Reflection/AssemblyLoader.cs
#	Kooboo.Sites/Repository/SiteRepositoryBase.cs

+++ Kooboo.Data/Infrastructure/InfraManager.cs
@@ -38,14 +38,13 @@ namespace Kooboo.Data.Infrastructure
             }
 
         }
-
-
-        public static bool Test(Guid OrganizationId, InfraType InfraType, int amount)
+                                                
+        public static bool Test(Guid OrganizationId, InfraType InfraType, long amount)
         {
             return instance.Test(OrganizationId, InfraType, amount); 
         }
 
-        public static void Add(Guid OrganizationId, InfraType InfraType, int amount, string message)
+        public static void Add(Guid OrganizationId, InfraType InfraType, long amount, string message)
         {
             instance.Add(OrganizationId, InfraType, amount, message); 
         }
